+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 ## Version 1
+### Version 1.1.0
+* New PlayerSkins module, including methods for getting player faces, fronts, and skins.
+* New BlockEngine module, including methods for getting 3D block images, 2D monster face images, and 3D full monster images.
+
 ### Version 1.0.1
 * Fix NoMethodError with get